requirements.txt
pyyaml==6.0
dbus-python==1.3.2
pycairo==1.21
PyGObject==3.42.2
